simple benchmarks for timing and time limiter
This is to get some comparisons for some of these operations
to see how much of a regression if any we are seeing due to
meltdown remediation. We see a high level impact on some
systems, this will hopefully help figure out what operations
we rely on might have gotten slower.